Initial commit of new interlock system, trimmed out useless code
