+++ Makefile
@@ -1,5 +1,5 @@
 # put your *.o targets here, make should handle the rest!
-SRCS = main.c system_stm32f0xx.c iolib.c serialio.c pwmout.c protos/messages.pb.c i2c.c reprog.c dripper.c hwaccess.c keycard.c ADClockout.c led_override.c
+SRCS = main.c system_stm32f0xx.c iolib.c serialio.c pwmout.c protos/messages.pb.c i2c.c reprog.c dripper.c hwaccess.c led_override.c interlock.c
 
 # all the files will be generated with this name (main.elf, main.bin, main.hex, etc)
 PROJ_NAME=main
