+++ inc/clock.h
@@ -1,6 +1,7 @@
 #ifndef __CLOCK_H
 #define __CLOCK_H
 
+void kick_watchdog();
 void delay_ms(int);
 
 #endif
