+++ inc/overrides.h
@@ -1,9 +1,8 @@
 	#ifndef INC_OVERRIDES_H_ 
 	#define INC_OVERRIDES_H_ 
  
-	#define ADC_KEY_EN 1 
 	#define INTERLOCK_KEY_EN 1 
-	#define LED_OVERRIDES_EN 0 
+	#define LED_OVERRIDES_EN 1 
 
 	#endif
