+++ overrides/not_safe_overrides.h
@@ -1,7 +1,6 @@
 	#ifndef INC_OVERRIDES_H_ 
 	#define INC_OVERRIDES_H_ 
  
-	#define ADC_KEY_EN 0 
 	#define INTERLOCK_KEY_EN 0 
 	#define LED_OVERRIDES_EN 1
